Minimize the number of operations we do in transactions
